rename comparisson -> comparison
